13 oct 1705

+++ frontend/controllers/PaymentController.php
@@ -2,6 +2,7 @@
 
 namespace frontend\controllers;
 
+use common\models\Admin;
 use common\models\Payment;
 use common\models\PaymentSearch;
 use common\models\Purpose;
@@ -12,14 +13,8 @@ use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\UploadedFile;
 
-/**
- * PaymentController implements the CRUD actions for Payment model.
- */
 class PaymentController extends Controller
 {
-    /**
-     * @inheritDoc
-     */
     public function behaviors()
     {
         return array_merge(
@@ -35,13 +30,12 @@ class PaymentController extends Controller
         );
     }
 
-    /**
-     * Lists all Payment models.
-     *
-     * @return string
-     */
     public function actionIndex()
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $searchModel = new PaymentSearch();
         $dataProvider = $searchModel->search($this->request->queryParams);
 
@@ -59,6 +53,10 @@ class PaymentController extends Controller
     }
 
     public function actionPay($id){
+        if(Yii::$app->user->isGuest){
+            return $this->redirect(['/site/login']);
+        }
+
         $teacher = Teacher::findOne(['user_id' => Yii::$app->user->identity->id]);
         $payment = new Payment();
         $payment->teacher_id = $teacher->id;
@@ -93,6 +91,10 @@ class PaymentController extends Controller
 
     public function actionReceipt($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $file = Payment::findOne($id);
 
         return Yii::$app->response->sendFile($file->payment, 'Квитанция.pdf', [
@@ -100,26 +102,23 @@ class PaymentController extends Controller
         ]);
     }
 
-    /**
-     * Displays a single Payment model.
-     * @param int $id ID
-     * @return string
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionView($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
     }
 
-    /**
-     * Creates a new Payment model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return string|\yii\web\Response
-     */
     public function actionCreate()
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $model = new Payment();
 
         if ($this->request->isPost) {
@@ -135,15 +134,12 @@ class PaymentController extends Controller
         ]);
     }
 
-    /**
-     * Updates an existing Payment model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param int $id ID
-     * @return string|\yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionUpdate($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $model = $this->findModel($id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -155,27 +151,17 @@ class PaymentController extends Controller
         ]);
     }
 
-    /**
-     * Deletes an existing Payment model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param int $id ID
-     * @return \yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionDelete($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
     }
 
-    /**
-     * Finds the Payment model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param int $id ID
-     * @return Payment the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     protected function findModel($id)
     {
         if (($model = Payment::findOne(['id' => $id])) !== null) {
