+++ frontend/controllers/ResultPdfController.php
@@ -2,20 +2,16 @@
 
 namespace frontend\controllers;
 
+use common\models\Admin;
 use common\models\ResultPdf;
 use common\models\ResultPdfSearch;
+use Yii;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
-/**
- * ResultPdfController implements the CRUD actions for ResultPdf model.
- */
 class ResultPdfController extends Controller
 {
-    /**
-     * @inheritDoc
-     */
     public function behaviors()
     {
         return array_merge(
@@ -31,13 +27,12 @@ class ResultPdfController extends Controller
         );
     }
 
-    /**
-     * Lists all ResultPdf models.
-     *
-     * @return string
-     */
     public function actionIndex()
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $searchModel = new ResultPdfSearch();
         $dataProvider = $searchModel->search($this->request->queryParams);
 
@@ -47,26 +42,23 @@ class ResultPdfController extends Controller
         ]);
     }
 
-    /**
-     * Displays a single ResultPdf model.
-     * @param int $id ID
-     * @return string
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionView($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
     }
 
-    /**
-     * Creates a new ResultPdf model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return string|\yii\web\Response
-     */
     public function actionCreate()
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $model = new ResultPdf();
 
         if ($this->request->isPost) {
@@ -82,15 +74,12 @@ class ResultPdfController extends Controller
         ]);
     }
 
-    /**
-     * Updates an existing ResultPdf model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param int $id ID
-     * @return string|\yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionUpdate($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $model = $this->findModel($id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -102,27 +91,17 @@ class ResultPdfController extends Controller
         ]);
     }
 
-    /**
-     * Deletes an existing ResultPdf model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param int $id ID
-     * @return \yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionDelete($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
     }
 
-    /**
-     * Finds the ResultPdf model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param int $id ID
-     * @return ResultPdf the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     protected function findModel($id)
     {
         if (($model = ResultPdf::findOne(['id' => $id])) !== null) {
