+++ frontend/controllers/TeacherController.php
@@ -2,6 +2,7 @@
 
 namespace frontend\controllers;
 
+use common\models\Admin;
 use common\models\Teacher;
 use common\models\TeacherSearch;
 use common\models\User;
@@ -11,14 +12,9 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
-/**
- * TeacherController implements the CRUD actions for Teacher model.
- */
 class TeacherController extends Controller
 {
-    /**
-     * @inheritDoc
-     */
+
     public function behaviors()
     {
         return array_merge(
@@ -34,13 +30,12 @@ class TeacherController extends Controller
         );
     }
 
-    /**
-     * Lists all Teacher models.
-     *
-     * @return string
-     */
     public function actionIndex()
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $searchModel = new TeacherSearch();
         $dataProvider = $searchModel->search($this->request->queryParams);
 
@@ -50,33 +45,31 @@ class TeacherController extends Controller
         ]);
     }
 
-    /**
-     * Displays a single Teacher model.
-     * @param int $id ID
-     * @return string
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionView($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
     }
 
-    /**
-     * Creates a new Teacher model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return string|\yii\web\Response
-     */
     public function actionCreate()
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $model = new SignupForm();
         $model2 = new Teacher();
         if ($model->load(Yii::$app->request->post())
             && $model2->load(Yii::$app->request->post())
             && $model->signup($model2)) {
 
-            return $this->redirect(['teacher/index']);
+            $model2->save(false);
+            return $this->redirect(['index']);
         }
 
         return $this->render('create', [
@@ -85,15 +78,12 @@ class TeacherController extends Controller
         ]);
     }
 
-    /**
-     * Updates an existing Teacher model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param int $id ID
-     * @return string|\yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionUpdate($id)
     {
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
         $model2 = $this->findModel($id);
         $model = User::findOne(['id' => $model2->user_id]);
 
@@ -116,16 +106,13 @@ class TeacherController extends Controller
         ]);
     }
 
-    /**
-     * Deletes an existing Teacher model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param int $id ID
-     * @return \yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionDelete($id)
     {
-        $teacher = Teacher::findOne($id);
+        if(Yii::$app->user->isGuest || !Admin::findOne(['user_id' => Yii::$app->user->identity->id])){
+            return $this->redirect(['/site/login']);
+        }
+
+        $teacher = $this->findModel($id);
         $user = User::findOne(['id' => $teacher->user_id]);
         $teacher->delete();
         $user->delete();
@@ -133,13 +120,6 @@ class TeacherController extends Controller
         return $this->redirect(['index']);
     }
 
-    /**
-     * Finds the Teacher model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param int $id ID
-     * @return Teacher the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     protected function findModel($id)
     {
         if (($model = Teacher::findOne(['id' => $id])) !== null) {
